server/b2b/requirements.txt
chromadb==0.5.0
fastapi==0.111.0
jose==1.0.0
langchain==0.1.17
langchain_community==0.0.36
langchain_core==0.1.50
motor==3.4.0
protobuf==5.26.1
pydantic==2.7.1
pydantic_core==2.18.2
pydantic_settings==2.2.1
pymongo==4.7.1
python-dotenv==1.0.1
python_jose==3.3.0
Requests==2.31.0
starlette==0.37.2
uvicorn==0.29.0
